pyproject.toml
[project]
name = "ml_pipelines_kfp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<3.11"
dependencies = [
    "fastapi>=0.111.1",
    "uvicorn>=0.30.1",
    "google-cloud-bigtable>=2.23.0",
    "google-cloud-secret-manager==2.20.1",
    "avro>=1.11.3",
    "fastavro>=1.9.5",
    "google-cloud-aiplatform>=1.59.0",
    "numpy>=2.0.0",
    "bugsnag>=4.7.1",
    "pytz>=2023.3",
    "google-cloud-logging>=3.10.0",
    "tenacity>=8.5.0",
    "pydantic>=2.6.4",
    "aiohttp>=3.9.3",
    "annotated-types>=0.7.0",
    "pydantic-core>=2.18.2",
    "click>=8.1.7",
    "h11>=0.14.0",
    "pathspec>=0.12.1",
    "xgboost>=2.1.1",
    "google-cloud>=0.34.0",
    "google-cloud-pipeline-components>=2.15.0",
    "google-cloud-storage>=2.16.0",
    "google-api-python-client>=2.137.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "protobuf>=3.17.0",
    "pandas>=2.2.2",
    "db-dtypes>=1.2.0",
    "optuna>=3.6.1",
    "scikit-learn>=1.5.1",
    "pytest",
    "pytest-asyncio>=0.23.8",
    "requests-mock>=1.11.0",
    "black",
    "ruff",
    "mypy",
    "httpx>=0.27.0",
    "coverage>=7.6.0",
    "whylogs==1.4.3",
    "polars>=0.20.29",
    "fugue[dask]>=0.9.0",
    "dask[dataframe,distributed]>=2024.6.2",
    "pandas-gbq>=0.23.0",
    "gcsfs>=2024.6.1",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.5",
    "kfp>=2.8.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"